% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{content}
\alias{content}
\alias{content.well}
\alias{content<-}
\alias{content<-.well}
\title{Get or set the contents of a well}
\usage{
content(x)

\method{content}{well}(x)

content(x) <- value

\method{content}{well}(x) <- value
}
\arguments{
\item{x}{A \code{well}}

\item{value}{A named list - what to supply the list}
}
\value{
For getting, a named list. For setting, a well.
}
\description{
Get or set the contents of a well
}
