% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{position}
\alias{position}
\alias{position<-}
\title{Get and set the position of a well}
\usage{
position(x)

position(x) <- value
}
\arguments{
\item{x}{A \code{well}}

\item{value}{A numeric vector}
}
\value{
For getting, a vector with two values. For setting, a well.
}
\description{
Get and set the position of a well
}
\details{
Coordinates follow the row, col indexing convention of dataframes
}
