% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{select_well}
\alias{select_well}
\alias{select_well.well}
\title{Select contents from a well}
\usage{
select_well(x, ...)

\method{select_well}{well}(x, select, ...)
}
\arguments{
\item{x}{the well to select}

\item{...}{arguments passed on to their respective methods}

\item{select}{Columns to select. Uses tidyselect.}
}
\value{
A \code{well}
}
\description{
Select contents from a well
}
