% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{update_well}
\alias{update_well}
\alias{update_well.well}
\title{Append to content in a well}
\usage{
update_well(x, ...)

\method{update_well}{well}(x, content_new, overwrite = TRUE, ...)
}
\arguments{
\item{x}{A well}

\item{...}{Arguments passed on to their repsective methods}

\item{content_new}{A named list of content to be added to the well}

\item{overwrite}{If multiple contents have the same name, should it be
overwritten?}
}
\value{
A \code{well}
}
\description{
Append to content in a well
}
