% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.R
\name{well}
\alias{well}
\title{Create a new well}
\usage{
well(x, y, content = list())
}
\arguments{
\item{x, y}{the x and y position of the well, where the bottom left hand well
is 1, 1}

\item{content}{a named list of key = value pairs containing well data}
}
\description{
Create a new well
}
\examples{
my_well <- well(
 1, 1,
 content =
   list(
     data = iris,
     dox = list(time = 1, time_unit = "hr", conc = 100, conc_unit = "ng/mL")
   )
)
}
